refactor: make rp-webhook a standalone command
